fix checkpoint deletion on cancel

+++ cloud/disk_manager/internal/pkg/services/common/checkpoints_for_images_and_snapshots.go
@@ -151,5 +151,9 @@ func CancelCheckpointCreation(
 		checkpointID = typedMetadata.CheckpointId
 	}
 
+	if checkpointID == "" {
+		return nil
+	}
+
 	return nbsClient.DeleteCheckpoint(ctx, disk.DiskId, checkpointID)
 }
